# dashboard/docs/generated/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "IT Global Dashboard API",
    "description": "IT Global 비즈니스 관리 대시보드를 위한 종합 API입니다.\n\n## 주요 기능\n- 프로젝트 관리 및 추적\n- 매출 및 재무 분석\n- 실시간 모니터링 및 메트릭\n- 사용자 관리 및 인증\n- 캐시 및 시스템 관리\n\n## 인증 방식\n이 API는 여러 인증 방법을 지원합니다:\n- 세션 기반 인증 (웹 인터페이스)\n- Bearer 토큰 인증 (API 접근)\n- Google OAuth 2.0 (웹 인터페이스)\n\n## 표준 응답 형식\n모든 API 응답은 다음과 같은 표준 형식을 따릅니다:\n```json\n{\n  \"success\": true|false,\n  \"data\": any,\n  \"error\": {...},\n  \"meta\": {...}\n}\n```\n\n## 버전 관리\nAPI는 URL 경로 기반 버전 관리를 사용합니다 (/api/v1/, /api/v2/ 등).",
    "version": "1.0.0",
    "contact": {
      "name": "IT Global Development Team",
      "email": "dev@itglobal.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://dashboard.itglobal.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/v1/docs/generate": {
      "post": {
        "operationId": "docs_api.generate_documentation",
        "summary": "API 문서 생성",
        "description": "summary: API 문서 자동 생성\ndescription: 현재 애플리케이션 상태를 분석하여 모든 형태의 API 문서를 생성합니다\nrequestBody:\ncontent:\napplication/json:\nschema:\ntype: object\nproperties:\nformats:\ntype: array\nitems:\ntype: string\nenum: [json, yaml, markdown, postman, guide]\ndescription: 생성할 문서 형식 (미지정시 모든 형식)\nforce:\ntype: boolean\ndescription: 강제 재생성 여부\ndefault: false\nresponses:\n200:\ndescription: 문서 생성 성공\n500:\ndescription: 문서 생성 실패",
        "tags": [
          "Documentation"
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "201": {
            "description": "생성됨",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionAuth": []
          }
        ]
      }
    },
    "/api/v1/docs/status": {
      "get": {
        "operationId": "docs_api.get_documentation_status",
        "summary": "문서 생성 상태 조회",
        "description": "summary: 문서 생성 시스템 상태 확인\ndescription: 현재 문서 생성 시스템의 상태와 통계 정보를 반환합니다\nresponses:\n200:\ndescription: 상태 조회 성공",
        "tags": [
          "Documentation"
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionAuth": []
          }
        ]
      }
    },
    "/api/v1/docs/validate": {
      "post": {
        "operationId": "docs_api.validate_documentation",
        "summary": "문서 유효성 검증",
        "description": "summary: 생성된 문서의 유효성 검증\ndescription: 현재 애플리케이션 상태와 생성된 문서의 일치성을 검증합니다\nresponses:\n200:\ndescription: 검증 완료",
        "tags": [
          "Documentation"
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "201": {
            "description": "생성됨",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionAuth": []
          }
        ]
      }
    },
    "/api/v1/docs/download/<file_type>": {
      "get": {
        "operationId": "docs_api.download_documentation",
        "summary": "문서 파일 다운로드",
        "description": "summary: 생성된 문서 파일 다운로드\ndescription: 지정된 형식의 문서 파일을 다운로드합니다\nparameters:\nin: path\nrequired: true\nschema:\ntype: string\nenum: [json, yaml, markdown, postman, guide, all]\nresponses:\n200:\ndescription: 파일 다운로드 성공\n404:\ndescription: 파일을 찾을 수 없음",
        "tags": [
          "Documentation",
          "name: file_type"
        ],
        "parameters": [
          {
            "name": "file_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "file_type 파라미터"
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionAuth": []
          }
        ]
      }
    },
    "/api/v1/docs/preview/<file_type>": {
      "get": {
        "operationId": "docs_api.preview_documentation",
        "summary": "문서 미리보기",
        "description": "summary: 생성된 문서 파일 미리보기\ndescription: 지정된 형식의 문서 내용을 JSON으로 반환합니다\nparameters:\nin: path\nrequired: true\nschema:\ntype: string\nenum: [json, yaml, markdown, postman, guide]\nresponses:\n200:\ndescription: 미리보기 성공\n404:\ndescription: 파일을 찾을 수 없음",
        "tags": [
          "Documentation",
          "name: file_type"
        ],
        "parameters": [
          {
            "name": "file_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "file_type 파라미터"
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionAuth": []
          }
        ]
      }
    },
    "/api/v1/docs/analytics": {
      "get": {
        "operationId": "docs_api.get_documentation_analytics",
        "summary": "문서 분석 정보 조회",
        "description": "summary: API 문서 분석 및 통계 정보\ndescription: 현재 API의 구조 분석 결과와 문서화 통계를 반환합니다\nresponses:\n200:\ndescription: 분석 정보 조회 성공",
        "tags": [
          "Documentation"
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionAuth": []
          }
        ]
      }
    },
...
    "/api/docs/": {
      "get": {
        "operationId": "flasgger.apidocs",
        "summary": "The /apidocs",
        "description": "/api/docs/ 경로에서 데이터를 조회합니다.",
        "tags": [
          "General"
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "StandardResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청 성공 여부"
          },
          "data": {
            "description": "응답 데이터"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetails"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "success",
          "data",
          "error",
          "meta"
        ]
      },
      "ErrorResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/StandardResponse"
          },
          {
            "type": "object",
            "properties": {
              "success": {
                "type": "boolean",
                "example": false
              },
              "data": {
                "type": "null"
              }
            }
          }
        ]
      },
      "ErrorDetails": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "에러 코드"
          },
          "message": {
            "type": "string",
            "description": "에러 메시지"
          },
          "details": {
            "type": "object",
            "description": "추가 에러 정보"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    }
  },
...
}